tidying up meta
Changing forked info for index, still have no idea how I'm going to create a general formula for this that is easily interpretable, wow this is going to be a confusing project for me. Also my thumb is broken so i wont be able to use two hands to code uber fast.